% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{read_blueprint}
\alias{read_blueprint}
\title{Read a Blueprint YAML File}
\usage{
read_blueprint(yaml, evaluate_r = TRUE)
}
\arguments{
\item{yaml}{Character. A path to a YAML file (.yaml or .yml) that contains a
valid blueprint for a compliant workbook.}

\item{evaluate_r}{Logical. Evaluate R code provided in your YAML file?
Defaults to `TRUE`.}
}
\value{
A nested list. The exact content and structure will depend on the
  supplied blueprint YAML file.
}
\description{
Read to a list a 'blueprint': a YAML file that encodes the content and
structure of a workbook compliant with the Analysis Function's best-practice
spreadsheet guidelines.
}
\details{
The YAML file must conform to a certain structure certain to be
  considered valid.

  R code can be inserted to the YAML file, prepended with `!expr`, and it
  will be evaluated on read if `evaluate_r` is `TRUE`. It's expected that
  you'll use this to read pre-created tables of data to be inserted into your
  workbook.
}
\examples{
# Read in the blueprint YAML file to a list
filepath <- system.file("extdata", "widgets.yaml", package = "yamlsheets")
blueprint <- read_blueprint(filepath)
blueprint

}
